examples/music-licensing-challenge/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
alembic==1.12.1 
strawberry-graphql==0.208.0
websockets==11.0.3